% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore}
\alias{baltimore}
\title{Baltimore neighborhood demographics}
\format{A data frame with 278 rows and 44 variables:
\describe{
\item{OBJECTID}{identifier for each neighborhood}
\item{Name}{name of the neighborhood}
\item{Population}{size of the population}
\item{White}{number of Caucasians}
\item{Blk_AfAm}{number of African Americans}
\item{...}{lots more}
}}
\usage{
baltimore
}
\description{
A dataset with demographic information for each of Baltimore's neighborhoods.
The data is available from the Baltimore City Open GIS Data \url{http://gis-baltimore.opendata.arcgis.com/}.
}
\examples{
data(baltimore_map)
data(baltimore)
baltimore_map <- dplyr::left_join(baltimore_map, baltimore, by="OBJECTID")
ggplot(baltimore_map, aes(x = long, y = lat, group = group)) +
 geom_polygon(aes(fill = Population))
ggplot(baltimore_map, aes(x = long, y = lat, group = group)) +
 geom_polygon(aes(fill = Blk_AfAm/Population*100))
}
\keyword{datasets}
