% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_map}
\alias{baltimore_map}
\title{Shape files for the neighborhoods of Baltimore}
\format{A data frame with 4732 rows and 6 variables:
\describe{
\item{OBJECTID}{identifier for each region}
\item{long}{geographic longitude}
\item{lat}{geographic latitude}
\item{group}{grouping variable for the polygon}
\item{order}{order of the rows}
\item{region}{name of the region}
}}
\usage{
baltimore_map
}
\description{
A dataset consisting of shapefiles describing the neighborhoods of Baltimore.
The shapefiles of the neighborhoods is available from the Baltimore City Open GIS Data \url{http://gis-baltimore.opendata.arcgis.com/}
}
\keyword{datasets}
